# book-network/src/main/java/com/jarviswuod/book/book/BookRequest.java
package com.jarviswuod.book.book;

import jakarta.validation.constraints.NotEmpty;
import jakarta.validation.constraints.NotNull;
import jakarta.validation.constraints.Size;

public record BookRequest(
        Long id,

        @NotNull(message = "100")
        @NotEmpty(message = "100")
        String title,

        @NotNull(message = "101")
        @NotEmpty(message = "101")
        String authorName,

        @NotNull(message = "102")
        @NotEmpty(message = "102")
        String isbn,

        @NotNull(message = "103")
        @NotEmpty(message = "103")
        @Size(max = 2000, message = "Synopsis cannot exceed 2000 characters")
        String synopsis,

        Boolean shareable
) {
    public BookRequest {
        if (shareable == null) {
            shareable = true;
        }
    }
}